# specs/001-piano-teacher/contracts/api.yaml
openapi: 3.0.3
info:
  title: Piano Teacher Management System API
  description: RESTful API for managing piano students, lessons, schedules, and homework
  version: 1.0.0
  contact:
    name: Piano Teacher System
    email: support@pianoteacher.com
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.pianoteacher.com/v1
    description: Production server
paths:
  /students:
    get:
      summary: Get all students
      description: Retrieve a list of all students with optional filtering
      tags:
        - Students
      parameters:
        - name: name
          in: query
          description: Filter by student name (partial match)
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size for pagination
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create new student
      description: Add a new student to the system
      tags:
        - Students
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /students/{id}:
    get:
      summary: Get student by ID
      description: Retrieve detailed information about a specific student
      tags:
        - Students
      parameters:
        - name: id
          in: path
          required: true
          description: Student ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: Update student
      description: Update student information
      tags:
        - Students
      parameters:
        - name: id
          in: path
          required: true
          description: Student ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: Delete student
      description: Remove a student from the system
      tags:
        - Students
      parameters:
        - name: id
          in: path
          required: true
          description: Student ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Student deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          maxLength: 100
          example: "张小明"
        age:
          type: integer
          minimum: 5
          maximum: 100
          example: 12
        grade:
          type: string
          maxLength: 50
          example: "六年级"
        email:
          type: string
          format: email
          example: "student@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+86-138-0013-8000"
        notes:
          type: string
          maxLength: 1000
          example: "对音乐有浓厚兴趣，练习认真"
        pricing:
          $ref: '#/components/schemas/Pricing'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateStudentRequest:
      type: object
      required:
        - name
        - age
      properties:
        name:
          type: string
          maxLength: 100
          example: "张小明"
        age:
          type: integer
          minimum: 5
          maximum: 100
          example: 12
        grade:
          type: string
          maxLength: 50
          example: "六年级"
        email:
          type: string
          format: email
          example: "student@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+86-138-0013-8000"
        notes:
          type: string
          maxLength: 1000
          example: "对音乐有浓厚兴趣，练习认真"
        pricing:
          $ref: '#/components/schemas/CreatePricingRequest'
    UpdateStudentRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: "张小明"
        age:
          type: integer
          minimum: 5
          maximum: 100
          example: 12
        grade:
          type: string
          maxLength: 50
          example: "六年级"
        email:
          type: string
          format: email
          example: "student@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+86-138-0013-8000"
        notes:
          type: string
          maxLength: 1000
          example: "对音乐有浓厚兴趣，练习认真"
    Pricing:
      type: object
      properties:
        id:
          type: integer
          format: int64
        hourlyRate:
          type: number
          format: decimal
          example: 150.00
        lessonDuration:
          type: number
          format: decimal
          example: 1.0
        paymentTerms:
          type: string
          maxLength: 500
          example: "月结"
    CreatePricingRequest:
      type: object
      required:
        - hourlyRate
        - lessonDuration
      properties:
        hourlyRate:
          type: number
          format: decimal
          minimum: 0
          example: 150.00
        lessonDuration:
          type: number
          format: decimal
          minimum: 0
          example: 1.0
        paymentTerms:
          type: string
          maxLength: 500
          example: "月结"
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...